Initial Release
This is the first public release. The zip file is the compiled
Macintosh application. The LiveCode file is the source code for the
compiled application which can be opened in the free Community version
of LiveCode available at www.livecode.com.

This version shows how to script calendar events in Calendar (formerly
called iCal) and Microsoft Outlook on the Macintosh. It was made using
Mac OS 10.10 (Yosemite) but should work on version 10.7 and later).